+++ tests/fixtures/external_plugins.py
@@ -33,12 +33,6 @@ string'
 '\''
 
 
-def complex_annotation(
-    first: List[Union[List[str], Dict[str, Dict[str, str]]]],
-):
-    ...
-
-
 def darglint_check(arg):
     """
     Used to trigger DAR101.