+++ tests/test_plugins.py
@@ -31,7 +31,6 @@ PLUGINS = (
     'S001',  # flake8-pep3101
     'S101',  # flake8-bandit
     'T100',  # flake8-debugger
-    'TAE002',  # flake8-annotations-complexity
     'RST215',  # flake8-rst-docstrings
     'EXE003',  # flake8-executable
     'DAR101',  # darglint