+++ wemake_python_styleguide/options/defaults.py
@@ -38,6 +38,12 @@ NESTED_CLASSES_WHITELIST: Final = (
     'Params',  # factoryboy specific
 )
 
+#: Domain names that are removed from variable names' blacklist.
+ALLOWED_DOMAIN_NAMES: Final = ()
+
+#: Domain names that extends variable names' blacklist.
+FORBIDDEN_DOMAIN_NAMES: Final = ()
+
 
 # ===========
 # Complexity:
@@ -112,8 +118,5 @@ MAX_COGNITIVE_AVERAGE: Final = 8  # based on this code statistics
 #: Maximum number of call chains.
 MAX_CALL_LEVEL: Final = 3
 
-#: Domain names that are removed from variable names' blacklist.
-ALLOWED_DOMAIN_NAMES: Final = ()
-
-#: Domain names that extends variable names' blacklist.
-FORBIDDEN_DOMAIN_NAMES: Final = ()
+#: Maximum number of nested annotations.
+MAX_ANN_COMPLEXITY: Final = 3